Refactor: model 패키지 위치 변경 / uistate 패키지 생성
