+++ app/src/main/java/com/konkuk/medicarecall/ui/model/CTAButtonType.kt
@@ -1,0 +1,5 @@
+package com.konkuk.medicarecall.ui.model
+
+enum class CTAButtonType {
+    WHITE, GREEN, DISABLED
+}
